--- src/app/dashboard/page.tsx
@@ -1,5 +1,0 @@
-import Dashboard from "@/components/user-dashboard/user-dashboard";
-
-export default function dashboardPage() {
-    return <Dashboard />;
-}
